refactor(invoice): extract available-stock calculation into shared util
Moves the cumulative-conversion-factor and available-stock-breakdown logic out of the invoice card into features/invoice/stock-availability.ts so other places (the new-invoice summary) can compute the same figures without duplicating the calculation.
